refactor: move some utils to validator folder
