+++ src/compare/subtract.ts
@@ -1,6 +1,6 @@
 import { Image, Mask } from '..';
-import checkProcessable from '../utils/checkProcessable';
-import { validateForComparison } from '../utils/validators';
+import checkProcessable from '../utils/validators/checkProcessable';
+import { validateForComparison } from '../utils/validators/validators';
 
 export interface SubtractImageOptions {
   /**
