+++ src/filters/autoLevel.ts
@@ -1,5 +1,5 @@
 import { Image } from '../Image';
-import checkProcessable from '../utils/checkProcessable';
+import checkProcessable from '../utils/validators/checkProcessable';
 import { ImageColorModel } from '../utils/constants/colorModels';
 
 export interface AutoLevelOptions {
