+++ src/load/decodePng.ts
@@ -1,7 +1,7 @@
 import { decode, DecodedPng } from 'fast-png';
 
 import { BitDepth, Image } from '../Image';
-import { assert } from '../utils/assert';
+import { assert } from '../utils/validators/assert';
 import { ImageColorModel } from '../utils/constants/colorModels';
 
 /**
