+++ src/morphology/morphologicalGradient.ts
@@ -1,7 +1,7 @@
 import { Image, Mask } from '..';
 import { subtract } from '../compare';
-import { checkKernel } from '../utils/checkKernel';
-import checkProcessable from '../utils/checkProcessable';
+import { checkKernel } from '../utils/validators/checkKernel';
+import checkProcessable from '../utils/validators/checkProcessable';
 
 export interface MorphologicalGradientOptions {
   /**
