+++ src/operations/thresholds/isodata.ts
@@ -5,7 +5,7 @@
  *
  */
 
-import { assert } from '../../utils/assert';
+import { assert } from '../../utils/validators/assert';
 
 /**
  * Return a threshold for a histogram using Isodata algorithm.
